+++ app/src/main/kotlin/com/robohorse/robopojogenerator/view/ViewModelMapper.kt
@@ -3,6 +3,7 @@ package com.robohorse.robopojogenerator.view
 import com.robohorse.robopojogenerator.form.GeneratorVew
 import com.robohorse.robopojogenerator.models.AdditionalPropertiesVM.Companion.GETTERS
 import com.robohorse.robopojogenerator.models.AdditionalPropertiesVM.Companion.JAVA_PRIMITIVE_TYPES
+import com.robohorse.robopojogenerator.models.AdditionalPropertiesVM.Companion.KOTLIN_DATA_CLASSES
 import com.robohorse.robopojogenerator.models.AdditionalPropertiesVM.Companion.KOTLIN_NULLABLE_FIELDS
 import com.robohorse.robopojogenerator.models.AdditionalPropertiesVM.Companion.KOTLIN_PARCELABLE
 import com.robohorse.robopojogenerator.models.AdditionalPropertiesVM.Companion.KOTLIN_SINGLE_DATA_CLASS
@@ -33,7 +34,8 @@ internal class ViewModelMapper(
             javaPrimitives = resolveCheckBox(JAVA_PRIMITIVE_TYPES),
             useLombokValue = resolveCheckBox(LOMBOK_VALUE),
             useTabsIndentation = useTabsIndentation.isSelected,
-            useMoshiAdapter = resolveCheckBox(MOSHI_ADAPTER)
+            useMoshiAdapter = resolveCheckBox(MOSHI_ADAPTER),
+            useKotlinDataClass = resolveCheckBox(KOTLIN_DATA_CLASSES)
         )
     }