+++ app/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 2.3.8
-- Moshi generate adapter option support.
+- Moshi: generate adapter option support.
+- Kotlin: regular classes support.
 - One character field name issue fix.
 
 ## 2.3.7